# Greenovation-backend-main/docs/openapiGreenovation.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Greenovation
  description: "This is the Open API documentation for the Greenovation API"
servers:
  - url: 'https://quason.tech/'
  - url: 'http://nattech.fib.upc.edu:40370/'
  - url: 'http://localhost:3000/'
paths:
  /users:
    post:
      tags:
        - Users
      summary: Creates a new user
      description: New User
      operationId: createUser
      requestBody:
        description: Provide the username, mail and password for the user
        required: true
        content:
          application/json:
            schema:
              required:
                - username
                - mail
                - password
              type: object
              properties:
                username:
                  type: string
                  example: 'My name'
                mail:
                  type: string
                  example: 'example@mail.com'
                password:
                  type: string
                  example: 'password1234'
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: 'Invalid parameters in body'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
    get:
      tags:
        - Users
      summary: Retrieves all Users
      operationId: getAllUsers
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/login:
    post:
      tags:
        - Users
      summary: Logs in a User
      operationId: loginUser
      requestBody:
        description: 'token can be null and user will not recieve notifications'
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: 'My name'
                password:
                  type: string
                  example: 'password1234'
                token:
                  type: string
                  example: null
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  accessToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoxLCJpYXQiOjE3MTMyNzI4MDMsImV4cCI6MTcxMzI3MzcwM30.b1kV_WYrckV0IzIZZQuSugQMCtTRNwW605sz7ojui1c'
                  refreshToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoxLCJpYXQiOjE3MTMyNzI4MDMsImV4cCI6MTcxMzI3NDAwM30.hyMTyFzyqqc9E3g5mdnk-BLi9uKqjfh_415xceiEUZ0'
        401:
          description: 'Unathorized to perform this operation'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        404:
          description: 'User with this username not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /auth/google:
    post:
      summary: Authenticate user with Google OAuth
      description: Authenticates a user using a Google OAuth token and returns access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: Google OAuth token
                  example: "your_google_oauth_token_here"
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: User ID
                    example: 123
                  accessToken:
                    type: string
                    description: Access token for authenticated user
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refreshToken:
                    type: string
                    description: Refresh token for authenticated user
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Invalid Google token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid Google token"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error"
  /users/refreshToken:
    post:
      tags:
        - Users
      summary: Refreshes tokens for the user
      operationId: refreshTokenUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  example: 1
                token:
                  type: string
                  description: 'This must be the refresh token you get from login or refresh'
                  example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoxLCJpYXQiOjE3MTMyNzI4MDMsImV4cCI6MTcxMzI3NDAwM30.hyMTyFzyqqc9E3g5mdnk-BLi9uKqjfh_415xceiEUZ0'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  accessToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoxLCJpYXQiOjE3MTMyNzI4MDMsImV4cCI6MTcxMzI3MzcwM30.b1kV_WYrckV0IzIZZQuSugQMCtTRNwW605sz7ojui1c'
                  refreshToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoxLCJpYXQiOjE3MTMyNzI4MDMsImV4cCI6MTcxMzI3NDAwM30.hyMTyFzyqqc9E3g5mdnk-BLi9uKqjfh_415xceiEUZ0'
        400:
          description: 'Invalid token'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
  /users/logout:
    delete:
      tags:
        - Users
      summary: Logs out refreshToken for the user
      operationId: logoutUser
      parameters:
        - name: token
          in: query
          description: 'This must be the refresh token you get from login or refresh'
          required: true
          schema:
            type: string
            example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoxLCJpYXQiOjE3MTMyNzI4MDMsImV4cCI6MTcxMzI3NDAwM30.hyMTyFzyqqc9E3g5mdnk-BLi9uKqjfh_415xceiEUZ0'
      responses:
        204:
          description: 'No content'
  /users/username/{username}:
    get:
      tags:
        - Users
      summary: Gets the user with the indicated username
      operationId: getUserByUsername
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: 'User with this username not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /users/mail/{mail}:
    get:
      tags:
        - Users
      summary: Gets the user with the indicated mail
      operationId: getUserByMail
      parameters:
        - name: mail
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: 'User with this mail not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Gets the user with the indicated id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: 'User with this id not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
    put:
      tags:
        - Users
      security:
        - bearerAuth: []
      summary: Edits the user with the indicated id
      operationId: putUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: 'New name'
                mail:
                  type: string
                  example: 'newmail@gmail.com'
                password:
                  type: string
                  example: '12345'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: 'Token not present'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        401:
          description: 'Unauthorized to edit this user'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Invalid token'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
        404:
          description: 'User with this id not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
    delete:
      tags:
        - Users
      security:
        - bearerAuth: []
      summary: Edits the user with the indicated id
      operationId: deleteUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: 'No content'
        401:
          description: 'Unauthorized to delete this user'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Invalid token'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
        404:
          description: 'User with this id not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /users/{id}/completableMissions:
    get:
      tags:
        - Users
        - Missions
      summary: Retrieves all completable missions by the user with the indicated id
      operationId: GetCompletableMissionsOfUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Missio'
        404:
          description: 'User with this id not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /users/{id}/missionsInProgress:
    get:
      tags:
        - Users
        - Missions
      summary: Retrieves all missions that the user can and has yet to complete
      operationId: GetMissionsInProgressOfUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Missio'
        404:
          description: 'User with this id not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /users/{id}/availableEvents:
    get:
      tags:
        - Users
        - Events
      summary: Retrieves all events associated to the user's missions in progress that he has yet to complete
      operationId: GetAvailableEventsOfUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        404:
          description: 'User with this id not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: 'My username'
        mail:
          type: string
          example: 'example@mail.com'
        password:
          type: string
          example: 'qMJt6gA^)V@e?:#x'
        diners:
          type: integer
          example: 15000
        avatar:
          type: integer
          example: 1
        idioma:
          type: string
          example: 'en'
        createdAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
        updatedAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
...
    Missio:
      type: object
      properties:
        id:
          type: number
          example: 1
        nameCat:
          type: string
          example: "Primera missió"
        nameEsp:
          type: string
          example: "Primera misión"
        nameEng:
          type: string
          example: "First mission"
        descriptionCat:
          type: string
          example: "Completa un esdeveniment qualsevol."
        descriptionEsp:
          type: string
          example: "Completa un evento cualquiera."
        descriptionEng:
          type: string
          example: "Complete any event."
        eventsForCompletion:
          type: number
          example: 1
        lletraId:
          type: number
          example: 7
        energy:
          type: number
          example: 20
        CO2:
          type: number
          example: 10
        happiness:
          type: number
          example: 100
        money:
          type: number
          example: 1000
        createdAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
        updatedAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
  examples:
...
    error400:
      value:
        message: 'Invalid information entered'
    error401:
      value:
        message: 'Unauthorized to perform this operation'
    error403:
      value:
        message: 'Invalid token'
    error404:
      value:
        message: 'Not found'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT